+++ src/odr/internal/html/filesystem.cpp
@@ -188,6 +188,7 @@ public:
     out.write_header_target("_blank");
     out.write_header_title("odr");
     write_viewport_meta(out, config(), false);
+    write_zoom_style(out, config(), false, {});
     write_filesystem_style(state);
     write_filesystem_dark_style(state);
     write_search_style(state);
@@ -267,6 +268,7 @@ public:
     out.write_element_end("table");
 
     write_search_script(state);
+    write_viewport_script(state);
 
     out.write_body_end();